Added docstring to SendEmail.py Refactored the sendEmail to SendEmail
